requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
gunicorn==21.2.0
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pillow==10.3.0
requests==2.31.0
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.18.1
google-cloud-vision==3.7.2
